+++ betty/plugins/entity/event.py
@@ -32,8 +32,8 @@ from betty.plugin.schema import PluginIdSchema
 from betty.plugins.entity.place import Place
 from betty.plugins.entity.presence import Presence
 from betty.plugins.event_type.unknown import Unknown as UnknownEventType
-from betty.plugins.role.subject import Subject
 from betty.privacy import Privacy
+from betty.roles.subject import Subject
 
 if TYPE_CHECKING:
     from collections.abc import Mapping
