+++ betty/privacy/privatizer.py
@@ -20,8 +20,8 @@ from betty.plugins.entity.place import Place
 from betty.plugins.entity.presence import Presence
 from betty.plugins.entity.source import Source
 from betty.plugins.event_type.death import Death
-from betty.plugins.role.subject import Subject
 from betty.privacy import Privacy
+from betty.roles.subject import Subject
 
 if TYPE_CHECKING:
     from collections.abc import Iterator, MutableSequence
