+++ betty/roles/attendee.py
