+++ betty/roles/speaker.py
