+++ betty/tests/plugins/asset/raspberry_mint/templates/entity/test_label__event.py
@@ -5,7 +5,7 @@ from betty.plugins.entity.person import Person
 from betty.plugins.entity.presence import Presence
 from betty.plugins.event_type.birth import Birth
 from betty.plugins.event_type.marriage import Marriage
-from betty.plugins.role.subject import Subject
+from betty.roles.subject import Subject
 from betty.test_utils.conftest import AssertTemplateFile
