+++ betty/tests/plugins/asset/raspberry_mint/templates/entity/test_summary__person.py
@@ -9,8 +9,8 @@ from betty.plugins.entity.presence import Presence
 from betty.plugins.entity.source import Source
 from betty.plugins.event_type.birth import Birth
 from betty.plugins.event_type.death import Death
-from betty.plugins.role.subject import Subject
 from betty.privacy import Privacy
+from betty.roles.subject import Subject
 from betty.test_utils.conftest import AssertTemplateFile
