+++ betty/tests/plugins/content/test_map.py
@@ -11,7 +11,7 @@ from betty.plugins.entity.event import Event
 from betty.plugins.entity.person import Person
 from betty.plugins.entity.place import Place
 from betty.plugins.entity.presence import Presence
-from betty.plugins.role.subject import Subject
+from betty.roles.subject import Subject
 from betty.test_utils.conftest import IsolatedProjectFactory
