+++ betty/tests/plugins/enricher/privatizer/test_jobs.py
@@ -12,8 +12,8 @@ from betty.plugins.entity.person import Person
 from betty.plugins.entity.presence import Presence
 from betty.plugins.entity.source import Source
 from betty.plugins.event_type.birth import Birth
-from betty.plugins.role.subject import Subject
 from betty.privacy import Privacy
+from betty.roles.subject import Subject
 
 if TYPE_CHECKING:
     from betty.test_utils.conftest import IsolatedProjectFactory
