+++ betty/tests/privacy/test_privatizer.py
@@ -18,11 +18,11 @@ from betty.plugins.entity.source import Source
 from betty.plugins.event_type.birth import Birth
 from betty.plugins.event_type.death import Death
 from betty.plugins.event_type.marriage import Marriage
-from betty.plugins.role.subject import Subject
-from betty.plugins.role.unknown import Unknown as UnknownRole
 from betty.privacy import Privacy
 from betty.privacy.privatizer import Privatizer
 from betty.project import DEFAULT_LIFETIME_THRESHOLD
+from betty.roles.subject import Subject
+from betty.roles.unknown import Unknown as UnknownRole
 from betty.test_utils.user import StaticUser
 
 if TYPE_CHECKING:
