+++ betty/tests/test_deriver.py
@@ -16,8 +16,8 @@ from betty.event_type import (
 from betty.plugins.entity.event import Event
 from betty.plugins.entity.person import Person
 from betty.plugins.entity.presence import Presence
-from betty.plugins.role.subject import Subject
 from betty.project import Project
+from betty.roles.subject import Subject
 from betty.test_utils.entity import record_added
 from betty.test_utils.locale.localizable import (
     DUMMY_COUNTABLE_LOCALIZABLE,
