+++ pyproject.toml
@@ -285,16 +285,16 @@ betty = 'betty.console:main_standalone'
 'village' = 'betty.plugins.place_type.village:Village'
 
 [project.entry-points.'betty.role']
-'attendee' = 'betty.plugins.role.attendee:Attendee'
-'beneficiary' = 'betty.plugins.role.beneficiary:Beneficiary'
-'celebrant' = 'betty.plugins.role.celebrant:Celebrant'
-'informant' = 'betty.plugins.role.informant:Informant'
-'organizer' = 'betty.plugins.role.organizer:Organizer'
+'attendee' = 'betty.roles.attendee:Attendee'
+'beneficiary' = 'betty.roles.beneficiary:Beneficiary'
+'celebrant' = 'betty.roles.celebrant:Celebrant'
+'informant' = 'betty.roles.informant:Informant'
+'organizer' = 'betty.roles.organizer:Organizer'
 'project-configuration' = 'betty.project._plugin_discovery:_discover_roles'
-'speaker' = 'betty.plugins.role.speaker:Speaker'
-'subject' = 'betty.plugins.role.subject:Subject'
-'unknown' = 'betty.plugins.role.unknown:Unknown'
-'witness' = 'betty.plugins.role.witness:Witness'
+'speaker' = 'betty.roles.speaker:Speaker'
+'subject' = 'betty.roles.subject:Subject'
+'unknown' = 'betty.roles.unknown:Unknown'
+'witness' = 'betty.roles.witness:Witness'
 
 [project.entry-points.'betty.http_rate_limit']
 'wikipedia-action-api' = 'betty.plugins.rate_limit.wikipedia_action_api:WIKIPEDIA_ACTION_API'
